requirements.txt
streamlit
groq
python-dotenv
requests
blessed
litellm